\name{PaulsTemplates-package}
\alias{PaulsTemplates-package}
\alias{PaulsTemplates}
\docType{package}
\title{Paul's Rmarkdown templates
}
\description{A group of templates for R markdown
}
\details{
\tabular{ll}{
Package: \tab PaulsTemplates\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-02-99\cr
License: \tab GPL3 \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{Paul Burch

Maintainer: Paul Burch <Paul.Burch77@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
